Generic command parameter validation and passing.
Use regex when validating the presence of the required parameters,
defined in the user job script.

Remove "%%{" from front and "%}" from back of the parameter value.

Don't escape command template paremeter value, if the command template
is generic and the parameter represents all the generic parameters.
The regex is used to check that the parameter value is indeed an escaped
string which can be passed to the Cluster.