Refactor FileTransferLogic and add debug log in ClusterAuthCredentialsRepo
Refactored the `FileTransferLogic.cs` file in the `HEAppE.BusinessLogicTier.logic.FileTransfer` namespace. The `SSHGenerator` object creation and public key generation have been moved inside an else block, which is executed only if the `ClusterUser`'s `AuthenticationType` is not `PrivateKeyInVaultAndInSshAgent`. If the `AuthenticationType` is `PrivateKeyInVaultAndInSshAgent`, a `FileTransferKeyCredentials` object is created with the `ClusterUser`'s credentials. The `FileTransferMethod` object creation has been split into two parts, with the `Credentials` property set inside the if-else block.

In the `ClusterAuthenticationCredentialsRepository.cs` file in the `HEAppE.DataAccessTier.Repository.UserAndLimitationManagement` namespace, a debug log statement has been added to log the decoded private key after importing vault data.