Refactor auth info retrieval and update logging and conditions
Refactored the code to retrieve authentication information from the `ClusterAuthenticationCredentialsRepository` using the `ClusterUser.Id` instead of directly from the `ClusterUser` object. Updated the logging statement to log the `authInfo` object instead of the `ClusterUser`. The condition to check the `AuthenticationType` has been updated to use pattern matching on the `authInfo` object. Also, the creation of the `FileTransferKeyCredentials` object now uses properties from the `credentials` object instead of the `ClusterUser` object.