Added vault data retrieval to ClusterProjectCredential
Added the `HEAppE.DataAccessTier.Vault` namespace to the import statements in `MiddlewareContext.cs`. In the `case ClusterProjectCredential clusterProjectCredentials:` block, new code was added to create a `VaultConnector` object and call its `GetClusterAuthenticationCredentials` method if the `entity` is of type `ClusterProjectCredential`. The returned data is then imported into the `ClusterAuthenticationCredentials` of the `entity` using the `ImportVaultData` method. This change allows for the retrieval and updating of data from a vault.