Simplify handling of ClusterProjectCredential objects
The code changes in `MiddlewareContext.cs` within the `HEAppE.DataAccessTier` namespace have been made to simplify the handling of `ClusterProjectCredential` objects. The original code was retrieving an entity from the database and then checking if it was of type `ClusterProjectCredential`. If it was, it was retrieving vault data and importing it into the entity's `ClusterAuthenticationCredentials`. The updated code simplifies this process by directly casting the retrieved entity and the item to `ClusterProjectCredential`. It then retrieves the vault data and imports it into the `ClusterAuthenticationCredentials` of the `clusterProjectCredentialEntity`. The `UpdateEntityOrAddItem` method call has been updated to use `entity` and `clusterProjectCredentialEntity` as arguments, suggesting that the method might have been updated to handle these specific types of arguments.