Update PublicKey assignment and refactor UserAndLimitationManagement
In `FileTransferLogic.cs`, the value assigned to `PublicKey` has been updated from `publicKey` to `credentials.PublicKeyFingerprint`. In `UserAndLimitationManagementLogic.cs`, the `_unitOfWork.Save();` line has been removed and replaced with a conditional return statement. If `hasUserGroup` is false, an `AuthenticationTypeException` is thrown with "NoUserGroup" and `user.Username` as arguments. If `hasUserGroup` is true, `user` is returned.