Add conditional check in FileTransferLogic.cs
Added a conditional check in the `FileTransferLogic.cs` file within the `HEAppE.BusinessLogicTier.logic.FileTransfer` namespace. The new code checks if the `AuthenticationType` of the `ClusterUser` in the `jobInfo.Specification` is of type `ClusterAuthenticationCredentialsAuthType.PrivateKeyInVaultAndInSshAgent`. If true, the function returns immediately, preventing further execution. This acts as a safeguard for cases where the authentication type is `PrivateKeyInVaultAndInSshAgent`.