feat: Update HashiCorp Vault initialization scripts
This commit updates the HashiCorp Vault initialization scripts. Specifically, it modifies the `initHVault.sh` script to convert `unseal_keys` into a JSON array and removes the unnecessary creation of `unseal_keys_json`. Additionally, it updates the `initHVaultConfig.sh` script to set the `ui` option to `false` instead of `true`. These changes improve the configuration and functionality of the HashiCorp Vault scripts.